[REFACTOR] move remaining 3dgui.hpp stuff
and include it directly where needed (instead of iengine.hpp)
