+++ inexor/engine/menus.cpp
@@ -1,6 +1,7 @@
 // menus.cpp: ingame menu system (also used for scores and serverlist)
 
 #include "inexor/engine/engine.hpp"
+#include "inexor/engine/3dgui.hpp"
 #include "inexor/ui.hpp"
 
 #define GUI_TITLE_COLOR  0xFFDD88
